% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estcure.R
\name{estcure}
\alias{estcure}
\title{Estimate cure rates for subgroups}
\usage{
estcure(object, z_value)
}
\arguments{
\item{object}{an object of fhcmodel.}

\item{z_value}{value(s) of the long-term covariate(s) in the order specified at the argument beta_variable of the fhcmodel().}
}
\value{
an estimated cure rate
}
\description{
Estimates a cure rate for a subgroup of patients with specified characteristics.
}
\examples{
result <- fhcmodel(data=fhc_dat,event_status = "event",event_time="time",,id="id",beta_variable = c("age","sex"),gamma_variable = c("age","sex"),se=F)
estcure(object=result,estvalue=c(0,1)) # cure rate for a subgroup population with age as the mean and female
estcure(object=result,estvalue=c(-20,0)) # cure rate for a subgroup population with 20 years younger than the mean age and male
}
