% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fhc_dat}
\alias{fhc_dat}
\title{A toy data set used for modeling the FHC model}
\format{
A data frame with 500 observations on the following 5 variables.
\describe{
\item{id}{patint id}
\item{age}{a continuous variable, which is centered to the mean}
\item{sex}{a binary variable with 0 for male, 1 for female}
\item{time}{observed survival time}
\item{event}{an event indicator with 1 for dead and 0 for censored}
}
}
\source{
{Generated from a FHC model to serve as an example.}
}
\usage{
fhc_dat
}
\description{
This toy data set has 500 patients' survival and baseline information and
contains a cure fraction.
}
\keyword{datasets}
