% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhcmodel.R
\name{fhcmodel}
\alias{fhcmodel}
\title{Fit Flexible-Hazards Cure Model}
\usage{
fhcmodel(
  data,
  event_time,
  event_status,
  id,
  beta_variable,
  gamma_variable = NULL,
  coef = T,
  se = T,
  max.int = 200,
  se.pert = 100,
  no_cores = 7,
  absconverge.par = 1e-06,
  relconverge.par = 1e-04,
  absconverge.F0t = 1e-06,
  relconverge.F0t = 0.001
)
}
\arguments{
\item{data}{a data.frame with no missing values contains survival time, event status, any continuous variable(s) and/or dummy variable(s) of categorical variable(s).
The categorical variables with more than 2 levels need to be converted into dummy variables before fitting the function.}

\item{event_time}{the name of observed survival time in the data.}

\item{event_status}{the name of event status in the data.}

\item{id}{the name of patient id in the data.}

\item{beta_variable}{the name(s) of variable(s) defined as long-term covariate(s) in the model, which cannot be NULL.}

\item{gamma_variable}{the name(s) of variable(s) defined as short-term covariate(s) in the model. By default gamma_variable = NULL.
If there is no short-term covariate that needs to be defined, gamma should be set as NULL and the model becomes a proportional hazards cure model.}

\item{coef}{a logical option for obtaining point estimates of regression parameters. By default coef = TRUE.}

\item{se}{a logical option for obtaining standard errors of regression parameters. By default se = TRUE.
If it is TRUE, the program returns standard error by perturbation method. If it is set to FALSE, the program only returns estimated regression parameters.}

\item{max.int}{maximum number of iterations. The default is 200.}

\item{se.pert}{the number of runs for perturbation. The default is 100.}

\item{no_cores}{the number of cores used while computing the standard errors by applying the perturbation. The default is 7.}

\item{absconverge.par}{absolute difference \eqn{(current-previous)} for regression parameters as the convergence criteria. The default is 1e-6.}

\item{relconverge.par}{relative difference \eqn{[(current-previous)/previous]} for regression parameters as the convergence criteria. The default is 1e-4.}

\item{absconverge.F0t}{the average of absolute differences for \eqn{F_0(t)} as the convergence criteria. The default is 1e-6.}

\item{relconverge.F0t}{the average of relative differences for \eqn{F_0(t)} as the convergence criteria. The default is 1e-3.}
}
\value{
a list containing results of the fit. The following items coef, iter, and data are returned when the argument coef=TRUE.
The item std.err and perturb are returned when the argument se=TRUE.
\item{coef}{estimated regression parameters (beta and gamma) if coef=TRUE}
\item{iter}{the number of iterations used to complete the point estimation if coef=TRUE}
\item{data}{the final data including the estimated \eqn{F_0(t)} and \eqn{f_0(t)} if coef=TRUE}
\item{std.err}{standard errors of regression paramters if se=TRUE}
\item{perturb}{a matrix containing the perturbation estimates, whose number of rows is se.pert, if se=TRUE}
}
\description{
Fits a flexible-hazards cure model. Tied failure times are dealt with Breslow approximation.
}
\examples{
data(fhc_dat)
result <- fhcmodel(data=fhc_dat,event_status = "event",event_time="time",id="id",
                   beta_variable = c("age","sex"),gamma_variable = c("age","sex"),se=F)
result <- fhcmodel(data=fhc_dat,event_status = "event",event_time="time",id="id",
                   beta_variable = c("age","sex"),se=T)
result$coef
}
